refactor(madories): improve ActionTabs quality and split tests
