+++ packages/madories/src/components/tool-sheet/action-tabs.tsx
@@ -1,4 +1,4 @@
-import { useState } from "react";
+import { useEffect, useState } from "react";
 import {
   Download,
   FolderOpen,
@@ -35,6 +35,7 @@ interface Props {
   onClear: () => void;
   onRotateFloor: () => void;
   onClose?: () => void;
+  defaultTab?: ActionTab;
 }
 
 export function ActionTabs({
@@ -50,25 +51,32 @@ export function ActionTabs({
   onClear,
   onRotateFloor,
   onClose,
+  defaultTab,
 }: Props) {
-  const [activeTab, setActiveTab] = useState<ActionTab>("edit");
+  const [activeTab, setActiveTab] = useState<ActionTab>(defaultTab ?? "edit");
   const [clearPending, setClearPending] = useState(false);
 
+  useEffect(() => {
+    if (!clearPending) return;
+    const id = setTimeout(() => setClearPending(false), 3000);
+    return () => clearTimeout(id);
+  }, [clearPending]);
+
   const editActions = [
-    { disabled: !canUndo, icon: <Undo2 size={14} />, onClick: onUndo, title: "戻す" },
-    { disabled: !canRedo, icon: <Redo2 size={14} />, onClick: onRedo, title: "進む" },
+    { disabled: !canUndo, icon: <Undo2 size={14} />, id: "undo", onClick: onUndo, title: "戻す" },
+    { disabled: !canRedo, icon: <Redo2 size={14} />, id: "redo", onClick: onRedo, title: "進む" },
   ];
 
   const fileActions = [
-    { disabled: false, icon: <Save size={14} />, onClick: () => { onSave(); onClose?.(); }, title: "保存" },
-    { disabled: false, icon: <FolderOpen size={14} />, onClick: () => { onLoad(); onClose?.(); }, title: "読込" },
-    { disabled: false, icon: <Download size={14} />, onClick: () => { onExportAll(); onClose?.(); }, title: "書出" },
-    { disabled: false, icon: <Link size={14} />, onClick: () => { onShare(); onClose?.(); }, title: "共有" },
+    { disabled: false, icon: <Save size={14} />, id: "save", onClick: () => { onSave(); onClose?.(); }, title: "保存" },
+    { disabled: false, icon: <FolderOpen size={14} />, id: "load", onClick: () => { onLoad(); onClose?.(); }, title: "読込" },
+    { disabled: false, icon: <Download size={14} />, id: "export", onClick: () => { onExportAll(); onClose?.(); }, title: "書出" },
+    { disabled: false, icon: <Link size={14} />, id: "share", onClick: () => { onShare(); onClose?.(); }, title: "共有" },
   ];
 
   const operationActions = [
-    { disabled: false, icon: <Maximize2 size={14} />, onClick: onFitView, title: "全体表示" },
-    { disabled: false, icon: <RotateCw size={14} />, onClick: () => { onRotateFloor(); onClose?.(); }, title: "回転" },
+    { disabled: false, icon: <Maximize2 size={14} />, id: "fitView", onClick: onFitView, title: "全体表示" },
+    { disabled: false, icon: <RotateCw size={14} />, id: "rotate", onClick: () => { onRotateFloor(); onClose?.(); }, title: "回転" },
   ];
 
   const tabDefs: { key: ActionTab; label: string; actions: typeof editActions }[] = [
@@ -85,6 +93,7 @@ export function ActionTabs({
         {tabDefs.map(({ key, label }) => (
           <button
             key={key}
+            aria-pressed={activeTab === key}
             style={{
               ...btnBase,
               background: activeTab === key ? "var(--ink)" : "transparent",
@@ -100,9 +109,10 @@ export function ActionTabs({
         ))}
       </div>
       <div style={{ display: "flex", gap: "4px" }}>
-        {currentActions.map(({ icon, onClick, title, disabled }) => (
+        {currentActions.map(({ icon, id, onClick, title, disabled }) => (
           <button
-            key={title}
+            key={id}
+            aria-label={title}
             title={title}
             disabled={disabled}
             onClick={onClick}
@@ -127,7 +137,6 @@ export function ActionTabs({
           onClick={() => {
             if (!clearPending) {
               setClearPending(true);
-              setTimeout(() => setClearPending(false), 3000);
             } else {
               setClearPending(false);
               onClear();
